requirements.txt
matplotlib==2.0.2
requests==2.17.3
python-telegram-bot==6.1.0
redis==2.10.5
python-dotenv==0.6.4
telegram-autoposter
